requirements.txt
flask==3.0.2
requests==2.32.2
gunicorn==22.0.0
PyYAML==6.0.1
stripe==7.11.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
